updated Post method for Groupmember controller
made post method check if the ids of customer and group exist before adding to database

+++ CustomerContactBook/Controllers/CustomersController.cs
@@ -36,7 +36,7 @@ namespace CustomerContactBook.Controllers
         /// </summary>
         /// <param name="id">customer id</param>
         [HttpGet("{id}")]
-        public async Task<ActionResult<Customer>> GetCustomer(int id)
+        public async Task<ActionResult<Customer>> GetCustomer(long id)
         {
             var customer = await _context.Customers.FindAsync(id);
